Move hash related files to foundation/hash
