+++ src/appleseed/foundation/hash/hash.h
