+++ src/appleseed/foundation/hash/murmurhash.cpp
