+++ src/appleseed/foundation/hash/murmurhash.h
