+++ src/appleseed/foundation/hash/siphash.cpp
