+++ src/appleseed/foundation/hash/siphash.h
