+++ src/appleseed/foundation/meta/tests/test_hash.cpp
@@ -27,6 +27,7 @@
 //
 
 // appleseed.foundation headers.
+#include "foundation/hash/hash.h"
 #include "foundation/image/color.h"
 #include "foundation/image/colormap.h"
 #include "foundation/image/colormapdata.h"
@@ -34,7 +35,6 @@
 #include "foundation/image/genericimagefilewriter.h"
 #include "foundation/image/image.h"
 #include "foundation/image/pixel.h"
-#include "foundation/math/hash.h"
 #include "foundation/math/rng/xoroshiro128plus.h"
 #include "foundation/math/scalar.h"
 #include "foundation/math/vector.h"
@@ -50,7 +50,7 @@
 
 using namespace foundation;
 
-TEST_SUITE(Foundation_Math_Hash)
+TEST_SUITE(Foundation_Hash_Hash)
 {
     struct HistogramFixture
     {
@@ -419,7 +419,7 @@ TEST_SUITE(Foundation_Math_Hash)
 
     static std::uint32_t boost_hash_combine(const std::uint32_t h1, const std::uint32_t h2)
     {
-        // See foundation::combine_hashes() in foundation/math/hash.h for the constant derivation.
+        // See foundation::combine_hashes() in foundation/hash/hash.h for the constant derivation.
         return h1 ^ (h2 + 0x9E3779B9ul + (h1 << 6) + (h1 >> 2));
     }
