+++ src/appleseed/renderer/kernel/intersection/assemblytree.cpp
@@ -46,6 +46,7 @@
 #include "renderer/utility/bbox.h"
 
 // appleseed.foundation headers.
+#include "foundation/hash/siphash.h"
 #include "foundation/math/beziercurve.h"
 #include "foundation/math/permutation.h"
 #include "foundation/math/ray.h"
@@ -56,7 +57,6 @@
 #include "foundation/platform/timers.h"
 #include "foundation/utility/foreach.h"
 #include "foundation/utility/lazy.h"
-#include "foundation/utility/siphash.h"
 #include "foundation/utility/statistics.h"
 #include "foundation/utility/string.h"
