+++ src/appleseed/renderer/kernel/rendering/final/uniformpixelrenderer.cpp
@@ -47,11 +47,11 @@
 
 // appleseed.foundation headers.
 #include "foundation/containers/dictionary.h"
+#include "foundation/hash/hash.h"
 #include "foundation/image/canvasproperties.h"
 #include "foundation/image/image.h"
 #include "foundation/math/aabb.h"
 #include "foundation/math/filtersamplingtable.h"
-#include "foundation/math/hash.h"
 #include "foundation/math/population.h"
 #include "foundation/math/scalar.h"
 #include "foundation/math/vector.h"
