+++ src/appleseed/renderer/kernel/rendering/generic/genericframerenderer.cpp
@@ -47,9 +47,9 @@
 // appleseed.foundation headers.
 #include "foundation/containers/dictionary.h"
 #include "foundation/core/concepts/noncopyable.h"
+#include "foundation/hash/hash.h"
 #include "foundation/image/canvasproperties.h"
 #include "foundation/image/image.h"
-#include "foundation/math/hash.h"
 #include "foundation/platform/thread.h"
 #include "foundation/utility/foreach.h"
 #include "foundation/utility/job.h"
