+++ src/appleseed/renderer/kernel/rendering/generic/generictilerenderer.cpp
@@ -44,11 +44,11 @@
 #include "renderer/utility/bbox.h"
 
 // appleseed.foundation headers.
+#include "foundation/hash/hash.h"
 #include "foundation/image/canvasproperties.h"
 #include "foundation/image/image.h"
 #include "foundation/image/tile.h"
 #include "foundation/math/aabb.h"
-#include "foundation/math/hash.h"
 #include "foundation/math/ordering.h"
 #include "foundation/math/scalar.h"
 #include "foundation/math/vector.h"
