+++ src/appleseed/renderer/modeling/input/texturesource.cpp
@@ -36,8 +36,8 @@
 #include "renderer/modeling/texture/texture.h"
 
 // appleseed.foundation headers.
+#include "foundation/hash/hash.h"
 #include "foundation/image/tile.h"
-#include "foundation/math/hash.h"
 #include "foundation/math/scalar.h"
 
 // Standard headers.
