+++ src/appleseed/renderer/modeling/surfaceshader/diagnosticsurfaceshader.cpp
@@ -51,11 +51,11 @@
 
 // appleseed.foundation headers.
 #include "foundation/containers/dictionary.h"
+#include "foundation/hash/hash.h"
 #include "foundation/image/color.h"
 #include "foundation/image/colorspace.h"
 #include "foundation/math/aabb.h"
 #include "foundation/math/distance.h"
-#include "foundation/math/hash.h"
 #include "foundation/math/minmax.h"
 #include "foundation/math/sampling/mappings.h"
 #include "foundation/math/scalar.h"
